requirements.txt
# Core Python
numpy
pandas

# Visualization
matplotlib
seaborn

# Jupyter / Notebooks
jupyter
ipykernel

# Text processing
nltk
regex

# Progress bars & utilities
tqdm

# Testing (already used by CI)
pytest



# Text processing & chunking
langchain
langchain-community
langchain-text-splitters

# Embeddings
sentence-transformers
torch

# Vector store
chromadb
faiss-cpu-1.13.2

